Refactor- fetch games with react query
Refactor UseGames.ts
