+++ src/hooks/useData.ts
@@ -1,46 +1,43 @@
-import React, { useState, useEffect } from "react";
-import apiClient from "../services/api-client";
-import { AxiosRequestConfig, CanceledError } from "axios";
 
-export type FetchResponse<T> = {
-  count: number;
-  results: T[];
-};
+// export type FetchResponse<T> = {
+//   count: number;
+//   results: T[];
+// };
 
-const useData = <T>(
-  endpoint: string,
-  requestConfig?: AxiosRequestConfig,
-  deps?: any[]
-) => {
-  const [data, setData] = useState<T[]>([]);
-  const [error, setError] = useState("");
-  const [isLoading, setLoading] = useState(false);
+// const useData = <T>(
+//   endpoint: string,
+//   requestConfig?: AxiosRequestConfig,
+//   deps?: any[]
+// ) => {
+//   const [data, setData] = useState<T[]>([]);
+//   const [error, setError] = useState("");
+//   const [isLoading, setLoading] = useState(false);
 
-  useEffect(
-    () => {
-      const controller = new AbortController();
+//   useEffect(
+//     () => {
+//       const controller = new AbortController();
 
-      setLoading(true);
-      apiClient
-        .get<FetchResponse<T>>(endpoint, {
-          signal: controller.signal,
-          ...requestConfig,
-        })
-        .then((res) => {
-          setData(res.data.results);
-          setLoading(false);
-        })
-        .catch((err) => {
-          if (err instanceof CanceledError) return;
-          setError(err.message);
-          setLoading(false);
-        });
-      return () => controller.abort();
-    },
-    deps ? [...deps] : []
-  );
+//       setLoading(true);
+//       apiClient
+//         .get<FetchResponse<T>>(endpoint, {
+//           signal: controller.signal,
+//           ...requestConfig,
+//         })
+//         .then((res) => {
+//           setData(res.data.results);
+//           setLoading(false);
+//         })
+//         .catch((err) => {
+//           if (err instanceof CanceledError) return;
+//           setError(err.message);
+//           setLoading(false);
+//         });
+//       return () => controller.abort();
+//     },
+//     deps ? [...deps] : []
+//   );
 
-  return { data, error, isLoading };
-};
+//   return { data, error, isLoading };
+// };
 
-export default useData;
+// export default useData;
