+++ src/hooks/useGames.ts
@@ -1,6 +1,6 @@
-import { UseQueryOptions, useQuery } from "@tanstack/react-query";
+import {  useQuery } from "@tanstack/react-query";
 import { GameQuery } from "../App";
-import  { FetchResponse } from "./useData";
+import { FetchResponse } from "../services/api-client";
 import apiClient from "../services/api-client";
 
 export type Platform = {
