+++ src/hooks/useGenres.ts
@@ -1,5 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
-import useData, { FetchResponse } from "./useData";
+import { FetchResponse } from "../services/api-client";
+
 import genres from "../data/genres";
 import apiClient from "../services/api-client";
