+++ src/hooks/usePlatforms.ts
@@ -1,5 +1,7 @@
 import { useQuery } from "@tanstack/react-query";
-import useData, { FetchResponse } from "./useData";
+// import useData, { FetchResponse } from "./useData";
+
+import { FetchResponse } from "../services/api-client";
 import apiClient from "../services/api-client";
 import platforms from "../data/platforms";
