+++ src/services/api-client.ts
@@ -1,5 +1,10 @@
 import axios from "axios";
 
+export type FetchResponse<T> = {
+  count: number;
+  results: T[];
+};
+
 export default axios.create({
   baseURL: "https://api.rawg.io/api",
   params: {
